+++ app/src/main/java/com/example/currencyflow/data/processContainers.kt
@@ -1,8 +1,9 @@
 package com.example.currencyflow.data
 
 import android.util.Log
+import java.math.BigDecimal
 
-fun processContainers(conversionsMap: Map<String, Double>, containers: List<C>) {
+fun processContainers(conversionsMap: Map<String, BigDecimal>, containers: List<C>) {
     for (container in containers) {
         val conversionValue = getConversionRate(conversionsMap, container.from.symbol, container.to.symbol)
         if (conversionValue != null) {
